healthcare-genai-rag/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
psycopg[binary]
alembic
python-multipart